fgfp load: support "basic" waypoint type
Load waypoints with <type>basic</type> from fgfp files,
provided they have valid coordinates.

In Flightgear, a "basic" waypoint is one for which only coordinates
should be interpreted. They are used as:
(1) custom waypoints
(2) fallback when other waypoint types can not be correctly interpreted
    (e.g., if Flightgear exports a flightplan with a navaid missing from
    its db, this navaid will become a "basic" waypoint).

Ideally, in case (1) the wpt should be loaded as WaypointType::USER,
but not in case (2). However it is not clear how to distinguish the two
usecases when loading the file.

This commit always sets WaypointType::USER.
The alternative would cause any custom waypoint to show as invalid.