fgfp export: save user waypoints with type "basic"
In fgfp files, <type>basic</type> denotes waypoints for which only
coordinates should be interpreted. It is appropriate for user waypoints.

Flightgear would also accept user waypoints exported with type "navaid",
but only because it falls back to interpreting the waypoint as "basic"
when it fails to find the ident.